ODR-225: pdu power pollers fail because of unknown OID description type
This bug is caused by 'power' is not defined in oidDescriptionQueryMap,
so it will report "Unknown OID description map type: power" when
self.nodeType is not 'sinetica' (IPI PDU).
Because oidDescriptionQueryMap and oidDescriptionMap is not used in PDU
power status poller at present, so remove it here to make clean.