Add update-lookups for redfish range discovery
Here is the scenario we are trying to accomplish (The user already has
the nodes powered on)
1. knowing the Idrac IP, we want to run the redfish discovery workflow.
This step will create the obm and catalog redfish. We can add a step
here to get the host mac addresses and populate the lookup table.
2. Since the node is discovered a pxe discovery shouldn't trigger
automatically but we should have a pxe discovery workflow that the user
can run against a discovered node to populate the catalog with pxe
cataloging data.
https://rackhd.atlassian.net/browse/RAC-6777